+++ src/ast/nodes/ObjectExpression.ts
@@ -3,12 +3,20 @@ import CallOptions from '../CallOptions';
 import ExecutionPathOptions from '../ExecutionPathOptions';
 import Identifier from './Identifier';
 import { ForEachReturnExpressionCallback, SomeReturnExpressionCallback } from './shared/Expression';
-import { isUnknownKey, objectMembers, ObjectPath, ObjectPathKey, UNKNOWN_KEY } from '../values';
+import {
+	objectMembers,
+	ObjectPath,
+	ObjectPathKey,
+	LiteralValueOrUnknown,
+	UNKNOWN_KEY,
+	UNKNOWN_VALUE
+} from '../values';
 import { Node, NodeBase } from './shared/Node';
 import { NodeType } from './NodeType';
 import { NodeRenderOptions, RenderOptions } from '../../utils/renderHelpers';
 import { BLANK } from '../../utils/blank';
 import MagicString from 'magic-string';
+import Literal from './Literal';
 
 const PROPERTY_KINDS_READ = ['init', 'get'];
 const PROPERTY_KINDS_WRITE = ['init', 'set'];
@@ -43,6 +51,17 @@ export default class ObjectExpression extends NodeBase {
 				);
 			}
 		}
+	}
+
+	getLiteralValueAtPath(path: ObjectPath): LiteralValueOrUnknown {
+		if (path.length === 0) return UNKNOWN_VALUE;
+
+		const { properties, hasCertainHit } = this.getPossiblePropertiesWithName(
+			path[0],
+			PROPERTY_KINDS_READ
+		);
+		if (!hasCertainHit || properties.length > 1) return UNKNOWN_VALUE;
+		return properties[0].getLiteralValueAtPath(path.slice(1));
 	}
 
 	hasEffectsWhenAccessedAtPath(path: ObjectPath, options: ExecutionPathOptions) {
@@ -84,7 +103,7 @@ export default class ObjectExpression extends NodeBase {
 	): boolean {
 		if (path.length === 0) return true;
 		const subPath = path[0];
-		if (path.length === 1 && !isUnknownKey(subPath) && objectMembers[subPath]) {
+		if (path.length === 1 && typeof subPath === 'string' && objectMembers[subPath]) {
 			return false;
 		}
 
@@ -135,7 +154,7 @@ export default class ObjectExpression extends NodeBase {
 	): boolean {
 		if (path.length === 0) return true;
 		const subPath = path[0];
-		if (path.length === 1 && !isUnknownKey(subPath) && objectMembers[subPath]) {
+		if (path.length === 1 && typeof subPath === 'string' && objectMembers[subPath]) {
 			return predicateFunction(options)(objectMembers[subPath].returns);
 		}
 
@@ -169,8 +188,18 @@ export default class ObjectExpression extends NodeBase {
 			const property = this.properties[index];
 			if (kinds.indexOf(property.kind) < 0) continue;
 			if (property.computed) {
-				properties.push(property);
-			} else if ((<Identifier>property.key).name === name) {
+				const value = property.key.getLiteralValueAtPath([]);
+				if (String(value) === name) {
+					properties.push(property);
+					hasCertainHit = true;
+					break;
+				} else if (value === UNKNOWN_VALUE) {
+					properties.push(property);
+				}
+			} else if (
+				(property.key instanceof Identifier && property.key.name === name) ||
+				(property.key instanceof Literal && property.key.value === name)
+			) {
 				properties.push(property);
 				hasCertainHit = true;
 				break;