+++ test/form/samples/dynamic-resolvable-if-statements/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'allows using variables to resolve conditionals'
+};