+++ test/form/samples/object-expression/computed-properties/main.js
@@ -1,0 +1,28 @@
+const x = {
+	"1": 1,
+	[2]: 2
+};
+
+if (x[1] < 2) {
+	console.log('kept');
+} else {
+	console.log('removed');
+}
+
+if (x["1"] < 2) {
+	console.log('kept');
+} else {
+	console.log('removed');
+}
+
+if (x[2] > 1) {
+	console.log('kept');
+} else {
+	console.log('removed');
+}
+
+if (x["2"] > 1) {
+	console.log('kept');
+} else {
+	console.log('removed');
+}