+++ test/form/samples/side-effect-p/_expected/amd.js
@@ -1,10 +1,8 @@
 define(function () { 'use strict';
 
-	var bool = true;
-
 	const hs = document.documentElement.style;
 
-	if ( bool ) {
+	{
 		hs.color = "#222";
 	}