+++ test/form/samples/side-effect-p/_expected/system.js
@@ -3,11 +3,9 @@ System.register([], function (exports, module) {
 	return {
 		execute: function () {
 
-			var bool = true;
-
 			const hs = document.documentElement.style;
 
-			if ( bool ) {
+			{
 				hs.color = "#222";
 			}