+++ test/form/samples/side-effect-p/_expected/umd.js
@@ -4,11 +4,9 @@
 	(factory());
 }(this, (function () { 'use strict';
 
-	var bool = true;
-
 	const hs = document.documentElement.style;
 
-	if ( bool ) {
+	{
 		hs.color = "#222";
 	}