+++ test/form/samples/template-literal/computed-properties/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'Resolve template literals in computed property keys'
+};