+++ test/form/samples/template-literal/simplify-conditionals/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'Template literals can be simplified when used in conditionals'
+};