+++ test/form/samples/treeshake-json/main.js
@@ -1,0 +1,11 @@
+import { production, nested } from './config.json';
+
+if (production) {
+	console.log('production');
+} else {
+	console.log('development');
+}
+
+if (nested.specialConfig !== 1) {
+	console.log('removed');
+}