+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.59.0
 *unreleased*
+* Tree-shake statically analysable dynamic conditionals ([#2167](https://github.com/rollup/rollup/pull/2167))
 * Do not emit empty chunks when code-splitting or empty files when preserving modules ([#2128](https://github.com/rollup/rollup/pull/2128))
 * Add `esm` format alias ([#2102](https://github.com/rollup/rollup/pull/2102))
 * Improve handling of external modules imported as both default and named imports ([#2136](https://github.com/rollup/rollup/pull/2136))